Multi Language Invoice Extractor Project/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb # for vector embedding